backend/requirements.txt
fastapi
uvicorn
gunicorn
pydantic-settings
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
